get-calendar-events/requirements.txt
functions-framework==3.*
firebase-admin==6.*
requests==2.*
google-cloud-logging==3.*
google-cloud-firestore==2.*
beautifulsoup4==4.*
lxml==5.*